quinta feira 08/05/2025

+++ admin/eliminar_hospede.php
@@ -1,11 +1,19 @@
 <?php
 require '../conexao.php';
 
-$id = $_GET['id'];
+$id = $_GET['id'] ?? null;
+
+if (!$id) {
+    header("Location: hospedes.php?erro=ID inválido");
+    exit;
+}
 
 $stmt = $conexao->prepare("DELETE FROM hospedes WHERE H_id_hospede=?");
 $stmt->bind_param("i", $id);
-$stmt->execute();
 
-header("Location: hospedes.php");
+if ($stmt->execute()) {
+    header("Location: hospedes.php?sucesso=Hóspede eliminado com sucesso");
+} else {
+    header("Location: hospedes.php?erro=Erro ao eliminar hóspede");
+}
 exit;
